buildVars, NVDATool: Moved AddonInfo class to NVDATool
