+++ site_scons/site_tools/NVDATool/typings.py
@@ -1,0 +1,18 @@
+from typing import TypedDict
+
+
+
+class AddonInfo(TypedDict):
+	addon_name: str
+	addon_summary: str
+	addon_description: str
+	addon_version: str
+	addon_author: str
+	addon_url: str|None
+	addon_sourceURL: str|None
+	addon_docFileName: str
+	addon_minimumNVDAVersion: str|None
+	addon_lastTestedNVDAVersion: str|None
+	addon_updateChannel: str|None
+	addon_license: str|None
+	addon_licenseURL: str|None
